requirements.txt
litellm
datasets>=3.3.2
tqdm
python-dotenv
loguru
aiohttp
huggingface_hub
setuptools
transformers>=4.48.3
aiofiles
polars
